requirements.txt
django>=5.1.0,<5.2.0
djangorestframework
pyyaml
requests
django-cors-headers
djangorestframework-simplejwt